requirements.txt
streamlit
torch
transformers
pdfplumber
pillow
